pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mas"
version = "0.1.0"
description = "Multi-Agent System - Production-scale, data-agnostic, LangGraph-based orchestration"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }

dependencies = [
    # --- Orchestration & LLM ---
    "langgraph>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "langchain-community>=0.3.0",

    # --- Embeddings & Retrieval ---
    "chromadb>=0.5.0",
    "sentence-transformers>=3.0.0",

    # --- Knowledge Graph ---
    "neo4j>=5.0.0",
    "networkx>=3.0",

    # --- Document Processing ---
    "PyMuPDF>=1.24.0",
    "pdfplumber>=0.11.0",
    "python-docx>=1.0.0",
    "Pillow>=10.0.0",

    # --- API & Server ---
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    "httpx>=0.27.0",

    # --- Data & Validation ---
    "pydantic>=2.0",
    "pydantic-settings>=2.0",

    # --- LLMOps & Observability ---
    "langfuse>=2.0.0",
    "opentelemetry-api>=1.25.0",
    "opentelemetry-sdk>=1.25.0",
    "prometheus-client>=0.20.0",

    # --- MCP ---
    "mcp>=1.0.0",

    # --- Utilities ---
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "typer>=0.12.0",
    "tenacity>=9.0.0",
    "structlog>=24.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0",
    "ruff>=0.5.0",
    "mypy>=1.10",
]

[project.scripts]
mas = "mas.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/mas"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
